requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
numpy>=1.26.0

# Database
SQLAlchemy>=2.0.23

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# PDF Generation
fpdf>=1.7.2
reportlab>=4.0.0

# Authentication
bcrypt>=4.0.0
passlib>=1.7.4
python-jose>=3.3.0

# Utils
python-dateutil>=2.8.0
pytz>=2023.3
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Monitoring
loguru>=0.7.0